requirements.txt
aiomysql==0.2.0
amqp==5.3.1
asgiref==3.8.1
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcrypt==4.3.0
billiard==4.2.1
blinker==1.7.0
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.6
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==44.0.2
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
geographiclib==2.0
geopy==2.4.1
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
kombu==5.5.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
mysqlclient==2.1.1
netifaces==0.11.0
oauthlib==3.2.0
packaging==25.0
prompt_toolkit==3.0.51
pycparser==2.22
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==5.4.1
requests==2.32.3
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.3
ssh-import-id==5.11
systemd-python==234
typing_extensions==4.12.2
tzdata==2025.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.4.0
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
zipp==1.0.0